project files/requirements.txt
tensorflow
numpy
pandas
opencv-python
matplotlib
streamlit
Pillow